requirements.txt
streamlit==1.37.0
joblib==1.4.2
pandas==1.5.3
numpy==1.24.4
matplotlib==3.8.0
scikitlearn==1.2.2
pillow==10.4.0   